+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `Image::build_par` that uses multithreaded compression algorithm for creating context directory archive. This method is behind `par-compression` feature flag.
 - Add `ImageListOptsBuilder::shared_size` parameter
 - Add `SystemDataUsageOpts` parameter to `Docker::data_usage`
+- `Container::stop` and `Container::restart` now take `ContainerStopOpts` and `ContainerRestartOpts` as paramaters respectively
 
 # 0.12.0
 - Fix some integer fields that could be negative but previously were a usize like `ImageSummary::containers`